feat: execute code on Apple Metal devices (#138)
git-subtree-dir: contrib/SPIRV-Cross
git-subtree-split: 39ce5b46de41f9fc77a2eedc842cadcffcb0659f

+++ include/athena/backend/llvm/runtime/BackendAccessor.h
@@ -38,7 +38,7 @@ public:
 
 private:
   T* mData;
-  std::vector<uint64_t> mShape;
+  std::vector<size_t> mShape;
 
   auto linearIndex(std::initializer_list<size_t> idx) -> size_t {
     std::vector<size_t> unwrappedIdx{idx};
